Rename TextWidget::getText -> getText32

+++ vvv3d/core/graphics/fonts/font.cpp
@@ -4,9 +4,9 @@
 #include "mgrfreetype.h"
 #include <map>
 #include <vector>
+#include <vvv3d/core/graphics/fonts/systemfonts.hpp>
 #include <vvv3d/core/graphics/textures/texturepacker.h>
 #include <vvv3d/utils/myutils.h>
-#include <vvv3d/core/graphics/fonts/systemfonts.hpp>
 
 namespace vvv3d {
 
@@ -21,10 +21,10 @@ const std::u32string characters = U"abcdefghijklmnopqrstuvwxyz"
 void convert8to32tex(const void* in, size_t width, size_t height, void* out)
 {
     const uint8_t* u8 = static_cast<const uint8_t*>(in);
-    uint8_t* u32      = static_cast<uint8_t*>(out);
+    uint8_t* u32 = static_cast<uint8_t*>(out);
     const size_t size = width * height;
     for (size_t i = 0; i < size; ++i)
-        for (size_t j      = 0; j < 4; ++j)
+        for (size_t j = 0; j < 4; ++j)
             u32[i * 4 + j] = (j == 3) ? u8[i] : 0xff;
 }
 
@@ -37,11 +37,11 @@ std::vector<uint8_t> convert8to32tex(const void* in, size_t width,
 
     for (size_t y = 0; y < height; ++y)
         for (size_t x = 0; x < width; ++x) {
-            const size_t inPixelIndex  = y * width + x;
+            const size_t inPixelIndex = y * width + x;
             const size_t outPixelIndex = (height - y - 1) * width + x;
-            for (size_t j                  = 0; j < 4; ++j)
+            for (size_t j = 0; j < 4; ++j)
                 ret[outPixelIndex * 4 + j] = 0xff;
-            ret[outPixelIndex * 4 + 3]     = u8[inPixelIndex];
+            ret[outPixelIndex * 4 + 3] = u8[inPixelIndex];
         }
     return ret;
 }
@@ -55,10 +55,10 @@ static std::vector<Glyph> loadGlyphes(FT_Face face,
         FT_Load_Glyph(face, glyph_index, FT_LOAD_DEFAULT);
         if (face->glyph->format != FT_GLYPH_FORMAT_BITMAP)
             FT_Render_Glyph(face->glyph, FT_RENDER_MODE_NORMAL);
-        const auto glyph   = face->glyph;
-        const auto& bm     = face->glyph->bitmap;
-        const auto width   = bm.width;
-        const auto height  = bm.rows;
+        const auto glyph = face->glyph;
+        const auto& bm = face->glyph->bitmap;
+        const auto width = bm.width;
+        const auto height = bm.rows;
         const auto xoffset = glyph->metrics.horiBearingX / 64;
         const auto yoffset = glyph->metrics.horiBearingY / 64 - height;
         const auto advance = glyph->metrics.horiAdvance / 64;
@@ -78,9 +78,9 @@ static std::vector<Glyph> loadGlyphes(FT_Face face,
 
 static bool glyphSizeComparisionHeightFirst(const Glyph& l, const Glyph& r)
 {
-    return l.height < r.height ? true :
-           l.height == r.height ? l.width < r.width
-                                : false;
+    return l.height < r.height
+               ? true
+               : l.height == r.height ? l.width < r.width : false;
 }
 
 static std::pair<uint32_t, uint32_t> glyphGetSize(const Glyph& g)
@@ -97,9 +97,9 @@ static void glyphSetOffset(Glyph& g, int32_t xoff, int32_t yoff, int32_t border)
 static void clearTexture(std::shared_ptr<Texture> lltex)
 {
     const size_t numPixels = lltex->getWidth() * lltex->getHeight() * 4;
-    void* ptr              = calloc(1, numPixels);
+    void* ptr = calloc(1, numPixels);
 
-    const auto width  = static_cast<GLsizei>(lltex->getWidth());
+    const auto width = static_cast<GLsizei>(lltex->getWidth());
     const auto height = static_cast<GLsizei>(lltex->getWidth());
     lltex->bind();
     glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, width, height, GL_RGBA,
@@ -145,7 +145,7 @@ FontImpl::FontImpl(FT_Face f, unsigned int size, unsigned int charSize,
     FT_Set_Char_Size(face, charSize * 64, charSize * 64, dpi, dpi);
     FT_Set_Pixel_Sizes(face, 0, size);
 
-    vector<Glyph> glyphes   = loadGlyphes(face, characters);
+    vector<Glyph> glyphes = loadGlyphes(face, characters);
     vector<Glyph> notPlaced = pack2d(glyphes, textureSize, textureSize,
                                      glyphSizeComparisionHeightFirst,
                                      glyphGetSize, glyphSetOffset, glyphes, 0);
@@ -161,22 +161,22 @@ void Font::activate(GLuint texUnit) { pImpl->lltex->bind(texUnit); }
 
 const Texture& Font::getTexture() const { return *pImpl->lltex; }
 
-long Font::getAscender() const
+int Font::getAscender() const
 {
-    return pImpl->face->size->metrics.ascender / 64;
+    return static_cast<int>(pImpl->face->size->metrics.ascender / 64);
 }
 
-long Font::getDescender() const
+int Font::getDescender() const
 {
-    return pImpl->face->size->metrics.descender / 64;
+    return static_cast<int>(pImpl->face->size->metrics.descender / 64);
 }
 
-long Font::getHeight() const
-{
-    return getAscender() - getDescender();
-}
+int Font::getHeight() const { return getAscender() - getDescender(); }
 
-long Font::getMinLeftGlyphEdge() const { return pImpl->face->bbox.xMin; }
+int Font::getMinLeftGlyphEdge() const
+{
+    return static_cast<int>(pImpl->face->bbox.xMin);
+}
 
 Font::Font() : pImpl() {}
 
@@ -202,7 +202,7 @@ void FontManager::addFont(const string& name, const string& filename,
                           unsigned int fontsize)
 {
     FT_Face face = freetypeMgr->addFont(name, filename);
-    auto f       = new Font();
+    auto f = new Font();
     f->pImpl.reset(new FontImpl(face, fontsize, 16, 96, 256));
     fonts[name].reset(f);
 }
@@ -216,4 +216,4 @@ const Font& FontManager::getFont(const string& name) const
 }
 
 FontManager::~FontManager() = default;
-}
+} // namespace vvv3d
