+++ vvv3d/core/graphics/fonts/font.h
@@ -27,10 +27,10 @@ public:
     const Glyph& getGlyph(uint32_t c) const;
     void activate(GLuint texUnit = 0);
     const Texture& getTexture() const;
-    long getAscender() const;
-    long getDescender() const;
-    long getMinLeftGlyphEdge() const;
-    long getHeight() const;
+    int getAscender() const;
+    int getDescender() const;
+    int getMinLeftGlyphEdge() const;
+    int getHeight() const;
 
 private:
     Font();
